052323 0103
Got the category of products to only load on the screen when that category's button was clicked.

minor format editing to make it easier to work with. All done in the CSS file and can be changed quickly to a new format that way

commented out two lines of code will go over with everyone Tuesday morning.